# src/main/resources/openapi/api.yaml
openapi: 3.0.3
info:
  title: Task Manager API
  description: Full-stack Task Manager REST API with categories, tags, and comments
  version: 1.0.0
  contact:
    name: Ankit Kumar Singh

servers:
  - url: http://localhost:8080
    description: Local development

tags:
  - name: Tasks
    description: Task CRUD operations
  - name: Categories
    description: Category CRUD operations
  - name: Tags
    description: Tag CRUD operations
  - name: Comments
    description: Task comment operations
  - name: Hello
    description: Health check / greeting endpoint

paths:
  # ===========================================================================
  # HELLO (existing endpoint — must stay working)
  # ===========================================================================
  /api/hello:
    get:
      tags:
        - Hello
      summary: Greeting endpoint
      operationId: sayHello
      responses:
        "200":
          description: Greeting message
          content:
            text/plain:
              schema:
                type: string
                example: Hello from Spring Boot CI/CD!

  # ===========================================================================
  # TASKS
  # ===========================================================================
  /api/tasks:
    get:
      tags:
        - Tasks
      summary: Get all tasks
      operationId: getAllTasks
      responses:
        "200":
          description: List of all tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaskResponse"

    post:
      tags:
        - Tasks
      summary: Create a new task
      operationId: createTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskRequest"
      responses:
        "201":
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"

  /api/tasks/{id}:
    get:
      tags:
        - Tasks
      summary: Get a task by ID
      operationId: getTaskById
      parameters:
        - $ref: "#/components/parameters/TaskId"
      responses:
        "200":
          description: Task found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskResponse"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      tags:
        - Tasks
      summary: Update a task
      operationId: updateTask
      parameters:
        - $ref: "#/components/parameters/TaskId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskRequest"
      responses:
        "200":
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      tags:
        - Tasks
      summary: Delete a task
      operationId: deleteTask
      parameters:
        - $ref: "#/components/parameters/TaskId"
      responses:
        "204":
          description: Task deleted successfully
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ===========================================================================
  # CATEGORIES
  # ===========================================================================
  /api/categories:
    get:
      tags:
        - Categories
      summary: Get all categories
      operationId: getAllCategories
      responses:
        "200":
          description: List of all categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryResponse"

    post:
      tags:
        - Categories
      summary: Create a new category
      operationId: createCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryRequest"
      responses:
        "201":
          description: Category created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"

  /api/categories/{id}:
    get:
      tags:
        - Categories
      summary: Get a category by ID
      operationId: getCategoryById
      parameters:
        - $ref: "#/components/parameters/CategoryId"
      responses:
        "200":
          description: Category found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryResponse"
        "404":
          description: Category not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      tags:
        - Categories
      summary: Update a category
      operationId: updateCategory
      parameters:
        - $ref: "#/components/parameters/CategoryId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryRequest"
      responses:
        "200":
          description: Category updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
        "404":
          description: Category not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      tags:
        - Categories
      summary: Delete a category
      operationId: deleteCategory
      parameters:
        - $ref: "#/components/parameters/CategoryId"
      responses:
        "204":
          description: Category deleted successfully
        "404":
          description: Category not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ===========================================================================
  # TAGS
  # ===========================================================================
  /api/tags:
    get:
      tags:
        - Tags
      summary: Get all tags
      operationId: getAllTags
      responses:
        "200":
          description: List of all tags
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TagResponse"

    post:
      tags:
        - Tags
      summary: Create a new tag
      operationId: createTag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TagRequest"
      responses:
        "201":
          description: Tag created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"

  /api/tags/{id}:
    get:
      tags:
        - Tags
      summary: Get a tag by ID
      operationId: getTagById
      parameters:
        - $ref: "#/components/parameters/TagId"
      responses:
        "200":
          description: Tag found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagResponse"
        "404":
          description: Tag not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      tags:
        - Tags
      summary: Update a tag
      operationId: updateTag
      parameters:
        - $ref: "#/components/parameters/TagId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TagRequest"
      responses:
        "200":
          description: Tag updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
        "404":
          description: Tag not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      tags:
        - Tags
      summary: Delete a tag
      operationId: deleteTag
      parameters:
        - $ref: "#/components/parameters/TagId"
      responses:
        "204":
          description: Tag deleted successfully
        "404":
          description: Tag not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ===========================================================================
  # COMMENTS (nested under tasks)
  # ===========================================================================
  /api/tasks/{taskId}/comments:
    get:
      tags:
        - Comments
      summary: Get all comments for a task
      operationId: getCommentsByTaskId
      parameters:
        - $ref: "#/components/parameters/CommentTaskId"
      responses:
        "200":
          description: List of comments for the task
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CommentResponse"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      tags:
        - Comments
      summary: Add a comment to a task
      operationId: addComment
      parameters:
        - $ref: "#/components/parameters/CommentTaskId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentRequest"
      responses:
        "201":
          description: Comment added successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommentResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/comments/{commentId}:
    delete:
      tags:
        - Comments
      summary: Delete a comment
      operationId: deleteComment
      parameters:
        - name: commentId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: Comment deleted successfully
        "404":
          description: Comment not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

# =============================================================================
# REUSABLE COMPONENTS
# =============================================================================
components:
  parameters:
    TaskId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64

    CategoryId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64

    TagId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64

    CommentTaskId:
      name: taskId
      in: path
      required: true
      schema:
        type: integer
        format: int64

  schemas:
    # =========================================================================
    # TASK
    # =========================================================================
    TaskStatus:
      type: string
      enum:
        - TODO
        - IN_PROGRESS
        - DONE

    TaskPriority:
      type: string
      enum:
        - LOW
        - MEDIUM
        - HIGH

    TaskRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          maxLength: 255
          example: Implement login page
        description:
          type: string
          maxLength: 1000
          example: Create the login form with email and password fields
        status:
          $ref: "#/components/schemas/TaskStatus"
        priority:
          $ref: "#/components/schemas/TaskPriority"
        dueDate:
          type: string
          format: date
          example: "2026-04-01"
        categoryId:
          type: integer
          format: int64
          description: Category ID to assign this task to
        tagIds:
          type: array
          uniqueItems: true
          items:
            type: integer
            format: int64
          description: Tag IDs to associate with this task

    TaskResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        status:
          $ref: "#/components/schemas/TaskStatus"
        priority:
          $ref: "#/components/schemas/TaskPriority"
        dueDate:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        category:
          $ref: "#/components/schemas/CategoryResponse"
        tags:
          type: array
          uniqueItems: true
          items:
            $ref: "#/components/schemas/TagResponse"
        commentCount:
          type: integer

    # =========================================================================
    # CATEGORY
    # =========================================================================
    CategoryRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          maxLength: 100
          example: Work
        color:
          type: string
          maxLength: 7
          example: "#FF5733"

    CategoryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        color:
          type: string
        createdAt:
          type: string
          format: date-time
        taskCount:
          type: integer

    # =========================================================================
    # TAG
    # =========================================================================
    TagRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          maxLength: 50
          example: urgent

    TagResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string

    # =========================================================================
    # COMMENT
    # =========================================================================
    CommentRequest:
      type: object
      required:
        - content
        - authorName
      properties:
        content:
          type: string
          maxLength: 2000
          example: This task needs clarification on the requirements
        authorName:
          type: string
          maxLength: 100
          example: Ankit

    CommentResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        content:
          type: string
        authorName:
          type: string
        createdAt:
          type: string
          format: date-time
        taskId:
          type: integer
          format: int64

    # =========================================================================
    # ERROR RESPONSES
    # =========================================================================
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
        status:
          type: integer
        error:
          type: string
        message:
          type: string

    ValidationError:
      type: object
      properties:
        timestamp:
          type: string
        status:
          type: integer
        error:
          type: string
        errors:
          type: object
          additionalProperties:
            type: string